Kak zhe ya ustal

+++ lab6/task_1/console.rb
@@ -3,6 +3,6 @@ require_relative 'main'
 
 puts 'Enter x'
 
-numbre = gets.to_i
+numbre = gets.to_f
 
 puts processing(numbre)
